# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status resolved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    post:
      tags: [Auth]
      operationId: loginWithDemoCredentials
      summary: Log in with a demo username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
        "401":
          description: Invalid credentials
...
  /approval-queue/{recordId}:
    post:
      tags: [Auth]
      operationId: reviewDailyRecord
      summary: Approve or return a submitted daily record
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Reviewed record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyOperations"
...
components:
  parameters:
...
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, username, email, firstName, lastName, profileImageUrl, role, department, isDemo]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [PRODUCTION_OPERATOR, QUALITY_OPERATOR, ENGINEERING_OPERATOR, STORES_OPERATOR, MANAGER, ADMIN]
        department:
          type: string
        isDemo:
          type: boolean
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    ReviewInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [START_REVIEW, APPROVE, REJECT]
        comments:
          type: string
...
    DailyOperations:
      type: object
      required: [id, factoryId, productionDate, season, shift, status, source, production, quality, efficiency, timeAccount, stoppages, energy, materials]
      properties:
        id:
          type: string
        factoryId:
          type: string
        productionDate:
          type: string
          format: date
        season:
          type: string
        shift:
          type: string
        status:
          type: string
          enum: [DRAFT, SUBMITTED, UNDER_REVIEW, APPROVED]
        source:
          type: string
        production:
          type: object
          additionalProperties: false
        quality:
          type: object
          additionalProperties: false
        efficiency:
          type: object
          additionalProperties: false
        timeAccount:
          type: object
          additionalProperties: false
        stoppages:
          type: array
          items:
            type: object
            additionalProperties: false
        energy:
          type: object
          additionalProperties: false
        materials:
          type: array
          items:
            type: object
            additionalProperties: false
...