# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: string;
  username: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role: AuthUserRole;
  department: string;
  isDemo: boolean;
}